requirements.txt
Flask
scikit-learn
nltk
PyPDF2
python-docx
numpy